Add ForeignAssets
